requirements.txt
numpy~=1.21.1
Pillow~=8.3.1
pywin32~=301
PyAutoGUI~=0.9.53
keyboard~=0.13.5
PyQt5~=5.15.4
jpype1